tests: break down tests into smaller modules
