--- src/tests/analyzers_test.py
@@ -1,189 +1,0 @@
-import ast
-import os
-from functools import partial
-
-from tryceratops import analyzers
-from tryceratops.violations import Violation, codes
-
-
-def read_sample(filename: str) -> ast.AST:
-    ref_dir = f"{os.path.dirname(__file__)}/samples/violations/"
-    path = f"{ref_dir}{filename}.py"
-
-    with open(path) as sample:
-        content = sample.read()
-        loaded = ast.parse(content)
-        return loaded
-
-
-def assert_violation(code: str, msg: str, line: int, col: int, violation: Violation):
-    assert violation.line == line
-    assert violation.col == col
-    assert violation.code == code
-    assert violation.description == msg
-
-
-def test_too_many_calls():
-    tree = read_sample("call_too_many_try")
-    analyzer = analyzers.call.CallTooManyAnalyzer()
-    expected_code, expected_msg = codes.TOO_MANY_TRY
-    assert_too_many = partial(assert_violation, expected_code, expected_msg)
-
-    violations = analyzer.check(tree, "filename")
-
-    assert len(violations) == 3
-    v2blocks, v3blocks_1, v3blocks_2 = violations
-
-    assert_too_many(15, 4, v2blocks)
-    assert_too_many(27, 4, v3blocks_1)
-    assert_too_many(32, 4, v3blocks_2)
-
-
-def test_raise_vanilla():
-    tree = read_sample("call_raise_vanilla")
-    analyzer = analyzers.call.CallRaiseVanillaAnalyzer()
-
-    assert_vanilla = partial(
-        assert_violation, codes.RAISE_VANILLA_CLASS[0], codes.RAISE_VANILLA_CLASS[1]
-    )
-
-    violations = analyzer.check(tree, "filename")
-
-    assert len(violations) == 1
-    violation = violations[0]
-
-    assert_vanilla(13, 8, violation)
-
-
-def test_raise_long_args():
-    tree = read_sample("call_raise_long_str")
-    analyzer = analyzers.call.CallRaiseLongArgsAnalyzer()
-
-    assert_args = partial(
-        assert_violation, codes.RAISE_VANILLA_ARGS[0], codes.RAISE_VANILLA_ARGS[1]
-    )
-
-    violations = analyzer.check(tree, "filename")
-
-    assert len(violations) == 1
-    violation, *_ = violations
-
-    assert_args(17, 8, violation)
-
-
-def test_check_continue():
-    tree = read_sample("call_check_continue")
-    analyzer = analyzers.call.CallAvoidCheckingToContinueAnalyzer()
-    msg = codes.CHECK_TO_CONTINUE[1].format("another_func")
-
-    assert_check = partial(assert_violation, codes.CHECK_TO_CONTINUE[0], msg)
-
-    violations = analyzer.check(tree, "filename")
-
-    assert len(violations) == 2
-    first, second = violations
-
-    assert_check(20, 4, first)
-    assert_check(24, 4, second)
-
-
-def test_reraise_no_cause():
-    tree = read_sample("except_reraise_no_cause")
-    analyzer = analyzers.exception_block.ExceptReraiseWithoutCauseAnalyzer()
-    code, msg = codes.RERAISE_NO_CAUSE
-
-    assert_no_cause = partial(assert_violation, code, msg)
-
-    violations = analyzer.check(tree, "filename")
-    assert len(violations) == 1
-
-    assert_no_cause(16, 8, violations[0])
-
-
-def test_verbose_reraise():
-    tree = read_sample("except_verbose_reraise")
-    analyzer = analyzers.exception_block.ExceptVerboseReraiseAnalyzer()
-    code, msg = codes.VERBOSE_RERAISE
-
-    assert_verbose = partial(assert_violation, code, msg)
-
-    violations = analyzer.check(tree, "filename")
-    assert len(violations) == 2
-    first, second = violations
-
-    assert_verbose(20, 8, first)
-    assert_verbose(28, 12, second)
-
-
-def test_broad_pass():
-    tree = read_sample("except_pass")
-    analyzer = analyzers.exception_block.ExceptBroadPassAnalyzer()
-    code, msg = codes.IGNORING_EXCEPTION
-
-    assert_broad = partial(assert_violation, code, msg)
-
-    violations = analyzer.check(tree, "filename")
-    assert len(violations) == 3
-    first, second, third = violations
-
-    assert_broad(18, 8, first)
-    assert_broad(27, 8, second)
-    assert_broad(35, 12, third)
-
-
-def test_consider_else():
-    tree = read_sample("try_consider_else")
-    analyzer = analyzers.try_block.TryConsiderElseAnalyzer()
-    code, msg = codes.CONSIDER_ELSE
-
-    assert_consider = partial(assert_violation, code, msg)
-
-    violations = analyzer.check(tree, "filename")
-    assert len(violations) == 1
-    violation = violations[0]
-
-    assert_consider(20, 8, violation)
-
-
-def test_try_inner_raise():
-    tree = read_sample("try_inner_raise")
-    analyzer = analyzers.try_block.TryShouldntRaiseAnalyzer()
-    code, msg = codes.RAISE_WITHIN_TRY
-
-    assert_inner_raise = partial(assert_violation, code, msg)
-
-    violations = analyzer.check(tree, "filename")
-    assert len(violations) == 2
-    violation = violations[0]
-
-    assert_inner_raise(21, 12, violation)
-
-
-def test_log_error():
-    tree = read_sample("log_error")
-    analyzer = analyzers.exception_block.LogErrorAnalyzer()
-    code, msg = codes.USE_LOGGING_EXCEPTION
-
-    assert_log_error = partial(assert_violation, code, msg)
-
-    violations = analyzer.check(tree, "filename")
-    assert len(violations) == 1
-    violation = violations[0]
-
-    assert_log_error(15, 8, violation)
-
-
-def test_log_object():
-    tree = read_sample("log_object")
-    analyzer = analyzers.exception_block.LogObjectAnalyzer()
-    code, msg = codes.VERBOSE_LOG_MESSAGE
-
-    assert_log_error = partial(assert_violation, code, msg)
-
-    violations = analyzer.check(tree, "filename")
-    assert len(violations) == 3
-    fstr, concat, comma = violations
-
-    assert_log_error(16, 40, fstr)
-    assert_log_error(23, 47, concat)
-    assert_log_error(30, 40, comma)
